requirements.txt
bottlenose==1.1.8
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.4
django-debug-toolbar==1.9.1
django-toolbelt==0.0.1
gunicorn==19.7.1
idna==2.6
lxml==4.2.1
pkg-resources==0.0.0
psycopg2==2.7.4
python-amazon-simple-product-api==2.2.11
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
urllib3==1.22
